Fix clippy.

+++ rs/execution_environment/benches/system_api/execute_inspect_message_wasm64.rs
@@ -29,7 +29,7 @@ pub fn execute_inspect_message_bench(c: &mut Criterion) {
             "ic0_msg_method_name_copy()/1B".into(),
             Module::InspectMessage.from_ic0(
                 "msg_method_name_copy",
-                Params3(0 as i64, 0 as i64, 1 as i64),
+                Params3(0_i64, 0_i64, 1_i64),
                 Result::No,
                 Wasm64::Enabled,
             ),
@@ -39,7 +39,7 @@ pub fn execute_inspect_message_bench(c: &mut Criterion) {
             "ic0_msg_method_name_copy()/30B".into(),
             Module::InspectMessage.from_ic0(
                 "msg_method_name_copy",
-                Params3(0 as i64, 0 as i64, 20 as i64),
+                Params3(0_i64, 0_i64, 20_i64),
                 Result::No,
                 Wasm64::Enabled,
             ),
